Added SQ4CD and ECO2DEMO
Decompilations of the CD version of SQ4 and the EcoQuest 2 demo are now available.

For some reason, SQ4 won't run at all, giving the error "2.v56 not found". When I make a dummy view with that number to circumvent this, it resulted in the game glitching up.
Even decompiling an unaltered ACTOR.SC and recompiling it (while leaving the other scripts unaltered) doesn't help; compiler bug?

Any ideas? Feel free to submit a pull request if you have any

On the other hand, ECO2DEMO has only a minor bug in rm130, where the bat's dialog overlaps its portrait (Adam's dialog displays properly). Since these messages are drawn entirely by the kernel, it's not an issue with any of the system scripts. Otherwise, the demo plays to the end with no known issues.