Added KQ6DEMO and PQ3DEMO
The decompilations for the demos of KQ6 and PQ3 are here.

Also, the Save script for LSL1VGA, LSL1DEMO, LSL5DEMO, and SQ1DEMO has been fixed to prevent the "duplicate saves" bug, using an original source script as the basis.

Now LSL1 plays to completion with no known issues.