LSL2, KQ4: All Main procedures identified
It turns out that the one unidentified procedure in LSL2 and KQ4 changes a sound object's loop and state. Thus, it is now known as ChangeSoundState, named after the DoSound parameter ChangeSndState. I wish I had a less confusing name for it.

Now all procedures in the Main script for KQ4 and LSL2 have been identified!